Deactivate the downstairs from the upstairs (#180)
Deactivate support for the Upstairs, part 1

These changes cover the first part of deactivate support in the
upstairs to allow it to cleanly disconnect from each downstairs,
sending a flush as a final command and waiting for its ACK.

Added an actual upstairs state instead of just an active boolean.
The supported states are Initializing, Active, and Deactivating.
Having specific states for the upstairs allows deactivate to work and
to tell the difference between initial startup connection, reconnects,
and deactivation.

Have the deactivation set method generate a flush to avoid a
race if the command before the deactivation was a flush.

More logic around activation and deactivation to make sure the
transitions happen at the correct time and report errors when a desired
change is not supported.

New crucible errors for deactivate situations.

Updated what happens when a downstairs tells the upstairs it is no
longer connected.

The process_message task now returns a handle that the main cmd_loop
can select on to detect if that task quits.

Updated cmd_loop() to know when a downstairs task has self
deactivated, and have cmd_loop also exit.

Much better logic in up_listen.  Just one loop now and better handling
of the differences between deactivation and just a single downstairs
going away and coming back.

Added checks for individual downstairs to determine if they can self
deactivate, when work plus a flush has finished.
Added a check to determine if all downstairs are deactivated and if
so, change the upstairs state to reflect that.

Renamed is_active() to guest_io_ready(). This is used to determine if
we are ready and accepting IO from a guest.

Deactivated state is now used for when downstairs completes the
work required for it to be deactivated.
Disabled is now the state for downstairs when they tell the upstairs
that a new upstairs has connected to them.

New BlockOp: Deactivate
The operation a guest sends to deactivate the upstairs from the
downstairs.

Added a bunch of tests for deactivation.

Created a new deactivate test for crucible-client
This test runs some IO, then does a deactivate then an activate.
Loops several times.  Verifies data from previous loops.
Added a "count" flag, currently only supported for deactivate test.
Added a retry loop for the client activate request with
a new option to the client test that will retry the initial client
activate request.  This helps the reconnect test, where it is
possible that downstairs going away and coming back can block the
activate longer than the default internal 10 retires that crucible currently
does.

Updated the test_reconnect.sh script to use the new client option to
retry forever.

Added deactivate test to test_up.sh script.

Co-authored-by: Alan Hanson <alan@oxide.computer>