Simple CLI for crucible client.
A super simple crucible-client server and cli program.
This is two programs based on the crucible-client test program.
A cli-server, which starts crucible-client, creates the runtime
and start the Upstairs.  This cli-sever opens a port and listens
for commands to come it, sending them to the upstairs (via the guest
methods).  Results are reported back to the CLI.

Some state is kept in the cli-server side.

The cli-client just takes a list of supported commands and sends
them over the socket to the server, reporting back their result.