Implement IOPS limiting Upstairs
Implement leaky bucket as a meter [1] to limit pulling block requests
from the guest's queue.

Also add "measure-iops" tool that borrows some ideas from fio: test a
random read-write workload with parameters for IO size, IO depth, IOPS
limit, and number of samples taken.

This commit also reworks the AWS benchmark to call this new tool, as
well as splits apart the scripts so that multiple benchmark runs can be
manually done (without editing driver.sh to add "echo" in the cleanup!).

Five runs of the "measure-iops" set to:

* IO depth = 1
* IO size = 1 block

and no IOPS limit shows (for the ubuntu test in ca-central-1):

    IOPS mean 491.22913 stddev 263.58258
    IOPS mean 492.43616 stddev 273.7362
    IOPS mean 507.68402 stddev 295.10162
    IOPS mean 507.74564 stddev 285.62772
    IOPS mean 499.91025 stddev 299.08554

With an IOPS limit of 400:

    IOPS mean 395.08057 stddev 76.68257
    IOPS mean 395.30582 stddev 78.66282
    IOPS mean 394.56558 stddev 82.57779
    IOPS mean 393.02597 stddev 80.48389
    IOPS mean 394.73245 stddev 86.41596

With an IOPS limit of 250:

    IOPS mean 246.34683 stddev 44.20688
    IOPS mean 246.91202 stddev 46.968605
    IOPS mean 247.36455 stddev 50.01197
    IOPS mean 247.9227 stddev 53.511436
    IOPS mean 248.18541 stddev 55.19138