New Dropshot server option to Upstairs to expose some stats.
Added an optional debug Dropshot server to the Upstairs that exposes a
few internal stats about Upstairs state and job queues.  This can grow
into more as we decide what other things we might want to look at.  This
is pretty much a copy of the Dropshot example code with some minor changes.

Fixed a few test warnings that crept in.