Implement snapshots, and update the crucible agent
Add ability to snapshot during flush by specifying an optional
"SnapshotDetails" struct, dictating that the Downstairs should take a
ZFS snapshot.

Teach the Downstairs how to serve from a read only filesystem, so it can
serve out of the ".zfs/snapshot/<name>" directory. A new "--mode"
argument can be "ro" or "rw" appropriately.

In the agent directory, update downstairs.xml for the change in how the
binary accepts arguments.

Update the agent. The agent now:

- correctly creates regions
- constructs ZFS datasets for the parent directory, and each region.
- knows to create regions with --encrypted as appropriate
- sets mode for downstairs SMF when booting snapshots
- can list snapshots, gets individual snapshot details, and delete
  snapshots
- can launch downstairs in read only mode for those snapshots, and
  delete the running snapshot downstairs as appropriate
- correctly deletes regions (have to disable downstairs smf instance
  first!)
- can accept X509 information in the region request to launch downstairs
  with TLS

Part of the agent update included changing the smf crate's Transaction
to hold on to TransactionEntries inside the Transaction itself.

This commit also adds the "serve" option to the downstairs: Josh
mentioned in one of the storage huddles that we may want to explore
running multiple regions off the same downstairs, so "serve" runs a
dropshot server that the agent can POST to run another instance of the
downstairs, instead of being required to run another SMF instance.

In upstairs, rename info.rs to http.rs, because now theres more
endpoints.