Downstairs reconcile part 2 (#202)
More work on the reconciliation between downstairs.

This next piece includes a specific function on the upstairs that runs
on initial startup that handles all the reconciliation work.  This new
function will determine where the mismatches are and how to fix them.
Once it has the work list, it converts it into messages it needs to
send to the downstairs to do that work. The work is sent to each
downstairs and the upstairs waits for an ACK from all three before
sending the next message.

The downstairs doing actual work is not written yet, that's coming soon.

The downstairs tasks all need to pass through a new function on their
way to active and this new function will be sure to do all the required
repair work before moving on to accepting IO from the guest.

Some new channels were created to communicate between the upstairs
reconcile task and the three downstairs repair tasks that do the work.

New fields to the Downstairs struct to store the work required to make
the three downstairs the same.

Rename a few things in mend.rs to better reflect their purpose.

Five new Message types to cover close/reopen of extents (and their
ack's) and an extent error message.

Updated the downstairs to be able to close and open individual extents.
Added a reopen all extents when an downstairs connects to an upstairs
to handle if an upstairs had disconnected and left a downstairs extent open.

Moved the stat_update function into a method on the upstairs so I could
call it while doing reconciliation.

The basic repair path is pretty straightforward.  Handling a downstairs or
and upstairs going away at any time, is as always, the real challenge.

Added more tests for the various reconcile methods.