Crucible Admin API updates
Updated the crucible admin endpoint to generate a openapi JSON file
that can then be used to create an actual admin client.  This is done
through a new upstairs binary that calls the function to display the API.

Added an admin-client directory for the admin-client library that
is generated using the openapi JSON file.  This library can be
used by programs wishing to access the Crucible Admin API.

Added a README in the openapi directory to describe how to create the
json file for crucible upstairs admin.

Added counters for completed and remaining repairs for upstairs
reconciliation. These counters are now part of the info included in
the upstairs stats structure.  Also added the state of each downstairs
to the upstairs stats structure.

Created a simple program (more coming on this later) that uses the
new Crucible Admin API to connect to an upstairs admin port and
request the info stats.  This program (cruadmin) is not part of the
Crucible workspace and has to be built manually.