Implement VolumeConstructionRequest
Because a volume can be arbitrarily deep, the intention is that Nexus
will store a JSON document in the database corresponding to a volume,
and POST that document to propolis when constructing a volume.

Implement VolumeConstructionRequest, a serializable/deserializable enum
that represents the volume to construct. All consumers (aka propolis)
have to do is pass this to Volume::construct, and they get a volume out
of it, and Nexus can store this in a JSONB column.