Compare read hash from different downstairs, test update
We currently send and receive reads from all three downstairs.
The first read is the result we send to the guest, the other two
are just dropped after they arrive in the upstairs.  This update
allows us to store and compare the hash from each read and report
if there is a difference.

When the reconciliation work finishes, this should be a panic or
some other serious error, as different read results should never
happen if things are working properly.

Added a bunch of tests for corruption that should also be updated
to panic or report serious error, when we decide what we should do.

Updated the test_up.sh to log if a test fails and print the name
of the failed tests at the end.