More client tests, other cleanup
Created a new fill test and a repair test.
The fill test just does a single write to every block in the
region one after the other.
The repair test does some random set of IOs, then exits as soon
as it can after getting an ack from all outstanding IOs.

Updated the logic around when to verify a volume, and by default
if you include a verify file it will not auto verify.  A new
verify option is enabled for that purpose.  Updated the verify
test to use this method.  By changing this, we can load a verify
file and not be required to also verify the entire volume.

Updated some tests to take the count option.

Fixed a clippy message from downstairs dump.

Updated the create-generic-ds.sh script to take block size as
an option.