New test CLI commands, updates for some tests
Added commands to run crucible-client fill and verify tests.
Added command to dump expected value for a block.
Updated the CLI server to support importing and exporting the
internal write count, and commands in the CLI to call them.

Updated fill and verify functions to write to 100 blocks at a
time.  Trying to do a single block at a time was taking forever
on larger regions.

Moved some cli logic outside the inner loop to allow for state
to be remembered across cli client reconnects.

Fixed the biggest workload test to still do something if the
region size was smaller than the largest IO supported.  Also updated
the biggest workload test to use the internal write counter and
update it accordingly.