Updates for Crucible client and CLI
Replace the crucible-client write_count with a new type that has a
bit more functionality.
Write log has the ability to remember a set write counts from a specific
time.  This allows is to have a range of write counts that a given block
could have.
This will allow creation of tests that will verify some failure/recovery
situations still have data that makes sense in terms of what could be
possible.

Added a commit cli command to force the write log to do a commit.
Added a bunch more tests around verifying that the write log is
doing the proper things in different situations.

Renamed some functions to better reflect their actions.