Perf test for crucible-client, other test tool updates
Created a new simple perf test in crucible-client.  Measure 100% writes
then 100% reads.  Caller can select io_size (blocks) and io_depth.

Updated create-generic-ds.sh to look for release binary before debug.

Updated downstairs-daemon.sh to use release binaries when requested.

Changed the Crucible QueryWorkQueue to return counts for both the
upstairs and the downstairs instead of just the upstairs.  This allows us
in a test to verify that all outstanding IOs have completed on all.
This is useful to prevent a previous test from interfering with a
test about to run.