test_up.sh: talk less, smile more
Update test_up.sh to send more output to a log file instead
of stdout.  This allows the test itself, and more so the
test when running in buildomat to run faster.

Log test output files to /tmp with .txt extension as this
is what buildomat archive will be looking for, when that change
comes in later.