Panic if read hash or decryption fails
This is part one of what we want to do if the upstairs finds
data from a read is not what it expects.  For now, we just
panic the upstairs.

Updated the tests to expect a panic in these situations.
A few were refactored and renamed to better reflect what they
are now testing.

Added a few more repair tests that walk all possible gen and flush
number differences.  This all should have been covered by the
existing tests, but I found them a little hard to follow to know
what exactly was being tested, and this new way I feel is much
more clear about what you are expecting in each case.