Panic if read hashes do not match between downstairs
When a read response comes back from a downstairs, we store
the hash that is returned.  When another downstairs returns
the same read request, we compare the hashes to make sure
that all downstairs are returning the same values.  A failure
of this check is now cause for panic.

However, during replay, there are situations where it's possible
to have a read return a different hash (and data) between the
downstairs.  When we are replaying, we take all jobs since the
last flush and replay them in order. This situation could result
in the contents of a read being different on a replaying
downstairs vs. the non-replaying downstairs if a write of the
same block came after the read.

Add a new field to the DownstairsIO struct to indicate when
an IO is a Replay IO and we can ignore a read hash mismatch
from other downstairs.

Other tidbits with this change

Updated show_all_work to display the replay field of the
DownstairsIO struct.

Connect up the cli show command to show the work queues.

Some new tests around making sure we panic on hash mismatch
except when under replay.

test_reconnect.sh improvements:  Change the test to do an
initial fill, then run the generic test during the loop.