crucible client perf and test_up updates (#304)
Added an option for the crucible-client perf test to save IO times to
a file in csv format.  This can (someday) allow for post processing.

Some formatting of comment strings to allow auto generated help
output to look better.  Changed a bunch of u32s to usize.

Update test_up.sh to send more output to a log file instead
of stdout.  This allows the test itself, and more so the
test when running in buildomat to run faster.

Log test output files to /tmp with .txt extension as this
is what buildomat archive will be looking for, when that change
comes in later.