Enable panic when hash mismatch (or failure) (#308)
Panic if read hash or decryption fails

This is part one of what we want to do if the upstairs finds
data from a read is not what it expects.  For now, we just
panic the upstairs.

Updated the tests to expect a panic in these situations.
A few were refactored and renamed to better reflect what they
are now testing.

Added a few more repair tests that walk all possible gen and flush
number differences.  This all should have been covered by the
existing tests, but I found them a little hard to follow to know
what exactly was being tested, and this new way I feel is much
more clear about what you are expecting in each case.

When a read response comes back from a downstairs, we store
the hash that is returned.  When another downstairs returns
the same read request, we compare the hashes to make sure
that all downstairs are returning the same values.  A failure
of this check is now cause for panic.

However, during replay, there are situations where it's possible
to have a read return a different hash (and data) between the
downstairs.  When we are replaying, we take all jobs since the
last flush and replay them in order. This situation could result
in the contents of a read being different on a replaying
downstairs vs. the non-replaying downstairs if a write of the
same block came after the read.

Add a new field to the DownstairsIO struct to indicate when
an IO is a Replay IO and we can ignore a read hash mismatch
from other downstairs.

Other tidbits with this change

Updated show_all_work to display the replay field of the
DownstairsIO struct.

Connect up the cli show command to show the work queues.

Some new tests around making sure we panic on hash mismatch
except when under replay.

test_reconnect.sh improvements:  Change the test to do an
initial fill, then run the generic test during the loop.

Co-authored-by: Alan Hanson <alan@oxide.computer>