Add a perf test to buildomat (#329)
Add a Perf test in CI

Created a new tools/test_perf.sh script to run the client
perf test on a variety of region sizes.  The results are
also written to CSV files.

Added this test_perf.sh script to the list of jobs buildomat
runs on release builds.

Updates to dsc for this test:
The start subcommand will now create a region, then spawn threads
to run the three downstairs, then just wait forever.  This will
grow to support more features soon, but for now it's a
replacement for "./tools/downstairs_daemon.sh -u".  Moving away
from bash scripts is the goal here, and this is the first step.
The start subcommand will now allow the caller to specify options
for creation of a region.

Updated the printed and CSV output for crucible-client perf test.

Co-authored-by: Alan Hanson <alan@oxide.computer>