Added a feature and grow the abilities of the dsc program.
dsc now supports just creating a region then exiting.

A dropshot server has been added to dsc, and now the "start" command
will start up three downstairs and a dropshot server.  This dropshot
server will take commands from outside and apply them to one or all
of the downstairs processes.  A bunch of internal plumbing was added
to support this including a work queue and several channels to
communicate between the tasks.

Cleaned up the names of some things, now that I know better what I'm
trying to do.

Added tools/test_dsc.sh to test the dropshot endpoints, but I believe
this will be short lived as coming next will be a real API user as
well as a test using the API and not curl.