flush_timeout, client perf, and downstairs DTrace
New CrucibleOpts field: flush_timeout
This adds the ability to select how long the Upstairs waits before
issuing
the internal flush. The ability to chose this at runtime will allow
tests to
try various values and report performance or other info without having
to recompile.

Client perf test updates:
Added read and write loop options to client perf test. This controls
how many of each type of IO workload we do.
Broke the read and write phases of the perf test into their own loops.

Added DTrace probes to the downstairs.
We now have probes for when work is taken off the wire, when work
is submitted to the OS, when work returns from the OS, and when work
is ACK'd back to the upstairs (put back on the wire).

This required passing the job_id down a layer deeper than it was before,
and adding job id's where none were needed before (and are ignored).

Added some DTrace example scripts.