Metric update for the Upstairs (#362)
This changes the way the upstairs handles metrics.  The new way is to
provide up_main() with a ProducerRegister handle that it can use to
register whatever stats it so desires.  The caller/user of up_main() is
responsible for creating a dropshot endpoint to serve the collection
requests coming from Oximeter.

Updated the client test to include a dropshot server and register with
Nexus if requested.

Removed the metric_collect and metric_register fields from CrucibleOpts
as they are now obsolete.  This required updating a few other files as well.

The Volume::construct() now also takes a ProducerRegister that it will
eventually pass to up_main().

Co-authored-by: Alan Hanson <alan@oxide.computer>