Multiple activations to read-only downstairs
In order to support snapshots, add support for a downstairs booted in
read-only mode to accept multiple read-only activations. This type of
activation only supports reads, and will create one Work struct for each
Upstairs that connects to it. Note that in order for this to not mix job
ids, *each Upstairs must have a different UUID when connecting*. This is
Nexus' responsibility.