Scrubber in the downstairs
Update downstairs writes to support a read_fill option.  This
option tells the extent to only do the write if the block has
not been written yet.  We can tell if the block was written by the
presence of a checksum.  Unwritten blocks have no checksum.

Added a bunch of tests to make sure writes of various sizes with
various written and unwritten blocks behave as expected.

Added some DTrace probes to cover the specific read_fill path.