dsc gets a shutdown endpoint
Added a new endpoint to dsc to allow remote shutdown of all
downstairs and then the dsc process itself.  This opens the door
to more bash script replacement.

Updated the communication channel between the dsc main task and the
sub-tasks for each downstairs to allow more than one outstanding
command at a time.

Updated the test_dsc.sh to test (and make use of) the /shutdown
endpoint.