Send session and gen number during activation
In the Downstairs, an Upstairs connection is now identified by its UUID,
a new session id (which is unique every time), and a generation number.

A connecting Upstairs will send these three things during the
negotiation (as part of HereIAm) and as part of activation (as part of
PromoteToActive). The Downstairs then has the option to accept or reject
based on more information. The Upstairs also receives more information
from the Downstairs when another Upstairs activates.

Note this commit *does not* add rejection from the Downstairs based on
connection generation number.

In order to support volumes where different nodes have different
generation numbers, there will probably be a future commit that removes
generation as an argument to `BlockIO::activate`.