Enforce read-only Upstairs and Downstairs (#398)
Move the `lossy` flag into the Upstairs, and add a `read_only` flag.
Make sure that an Upstairs in read-only mode cannot submit writes.

Supporting configuring Downstairs in read-only mode, which opens Extents
in read-only mode. Return errors if an Upstairs tries to modify on-disk
state.

Negotiation now sends read-only and encryption expectation, and the
Downstairs will reject connections where there is a mismatch.

Connections are closed if one of those mismatches are seen: there's no
way that the negotiation could succeed. This currently causes a loop
where the Upstairs attempts to reconnect indefinitely.