Remember to draw the rest of the owl
When Volumes have a read-only parent, `Volume::read` will conditionally
send two reads out in cases where the LBA range coverage overlaps: one
to the read-only parent, and one to the appropriate subvolume.

In the case the where the subvolume's coverage extends past the
read-only parent's coverage:

    read-only parent: AAAA
           subvolume: BBBBCCCC

There was a bug where blocks CCCC from the subvolume were not being
written into the buffer that is returned by `Volume::read`! This commit
fixes that, and adds a lot of tests to cover this case.

All credit goes to Alan (@leftwo) for finding this!