Support multiple read-only activations Downstairs (#411)
Previously, the Downstairs would only accept one active Upstairs at a
time. This commit adds support Downstairs for multiple read-only
Upstairs connecting, and sending reads.

This was done by storing a Work struct per Upstairs connection, and
getting the Downstairs to select the correct one based on an
UpstairsConnection argument.

The biggest change in this commit is to `promote_to_active`, which has
different behaviour now depending on if the Downstairs is in read-only
or read-write mode. The read-write path should be functionally the same
as before this commit. Tests were also added to validate
`promote_to_active`'s new behaviour.

In order to protect accesses to `Downstairs::active_upstairs`, functions
that read or modify it are now required to be through an exclusive
reference.