More features for dsc (#417)
* More features for dsc

Dsc can now support random restart of one of the downstairs under
its control. The requester can pick a min and max time to fence
the restarts. This required a little refactor of the dsc main loop
to allow for some persistent state.

Updated dsc to support sending a single command to a running dsc
server, basically making it a client of itself. To support this,
further refactoring of dsc was required.

Added the new dsc commands to the crutest cli.

Wrote a small test function to create a temporary file instead of
creating an empty file in the current directory which was not
removed at the end of the tests.

Updated the top level README.md to make it match the current
names of things.