Region repair perf test, updates for dsc
This is an initial version of an extent repair performance test.  We create
three downstairs, then make an entire downstairs set of extents require
repair.  We use output from the upstairs to determine how long that repair
takes, and divide by the number of extents to get an average.

Dsc has been updated to support each downstairs having its own path to the
region.  This allows downstairs to run on different physical disks if
desired.  Making this change required updating a number of places inside
dsc to support what is now a Vec.  There were already a bunch of assumptions
inside dsc on how the regions were setup.  This reduces that somewhat, but
there is still more we could do.

A small update to the upstairs that reports time spent in repair and
computation of the average repair time.