Implement missing "snapshot" integration test (#445)
There were two tests in upstairs/src/volume.rs that remained
unimplemented due to requiring a running set of downstairs:

        construct_snapshot_backed_vol
        construct_read_only_iso_volume

We have a way to write integration tests now, so this commit implements
a test for a "snapshot" backed volume. It does not perform a snapshot
but simply boots a downstairs in read-only mode and uses that as a
read-only parent.

In order to implement this, I wanted to write some data into a set of
downstairs, then boot new downstairs read-only but pointing at the same
on-disk files. The integration test framework couldn't support this
without adding `fn deactivate` to the BlockIO trait - the original
read-write downstairs have to be deactivated (and have their tasks
stopped) in order for the read-only downstairs to boot ok.

There's also some misc integration framework improvements:

- don't drop the tempdirs
- uniquely prefix each test's tempdir
- panic on port collisions

Closes https://github.com/oxidecomputer/crucible/issues/298