slog in the downstairs
This adds slog logging to the downstairs and replaces most of the
println calls with slog.

I chose to add a Logger field to the region structure as well as to
the Downstairs structure.  This resulted in the least amount of
having to pass around the Logger element.

build_downstairs_for_region uses an Optional log because the
downstairs admin dropshot endpoint can't take a log arg through the same
way it takes other args.

A smattering of tests also need the Logger element, so there are a few
csl functions (create simple logger) that is used to satisfy the need
for a Logger.

I removed a few println's that were really there just for debugging.

For the upstairs, there were still some println's in the scrubber code
that I did not deal with the last time.  The scrub method now requires
a Logger so it knows where to send its messages.