Bite the bullet: make BlockIO async (#449)
BlockIO is now an async trait. This was mostly a rote process of adding
async to every fn, and adding await everywhere.

BlockReqWaiter is now gone from the BlockIO trait. Consumers of
implementations of that trait now simply need to await on the result of
operations. BlockIO implementations are now expected to act like NVMe or
Virtio disks: operations are not expected to be processed in any order.

Consumers now await on the future(s) returned by BlockIO trait
functions, or can collect a list of futures and call `join_all` to wait
on all of them concurrently.

block_on and block_in_place should no longer be required. If you see
block_in_place now, it should be scrutinized and is probably wrong.
Likewise, if you're using std::sync::Mutex or other blocking
implementations from std, they should be changed to tokio equivalents.
Any std::thread::sleep use is a now a bug!

One place where block_in_place is still required is CruciblePseudoFile,
which implements std::io::{Read, Write, Seek}, which are *not* async.
Another is in the implementation for oximeter::Producer, which is not an
async trait.

Some misc things:

- extent_from_offset no longer returns Result, because it didn't needto

- FileBlockIO can use reqwest::Client instead of
  reqwest::blocking::Client, which is nice.

- Volume::add_read_only_parent_create_guest was removed, it wasn't being
  used